Refactor refactor refactor
