+++ src/client.rs
@@ -4,9 +4,9 @@ use std::fmt;
 use tokio::task::JoinHandle;
 use tokio::net::TcpStream;
 use tokio_native_tls::{TlsConnector, native_tls};
-use hyper::body::{Bytes, HttpBody};
 use hyper::{Body, Request, Response, StatusCode};
 use hyper::client::conn::{Builder, SendRequest};
+use select::document::Document;
 
 use crate::detour::Detour;
 
@@ -65,26 +65,23 @@ impl Client {
         self.sender.send_request(req)
     }
 
-    // send a GET and return only body. head is ignored
-    // does this absolutely needed?
-    pub fn get_bytes(&mut self, path: &str)
-        -> impl Future<Output = Result<Vec<u8>, Box<dyn Error>>> + '_ {
+    pub fn query_html(&mut self, path: &str)
+        -> impl Future<Output = Result<Document, Box<dyn Error>>> + '_ {
         let path = path.to_owned();
 
         async move {
             let req = Request::builder()
                 .method("GET")
+                .header("Host", self.host.as_str())
                 .uri(path)
-                .header("Host", &self.host)
-                .header("Accept", "*/*")
-                .body(Body::from(""))?;
+                .header("Accept", "text/html")
+                .body(Body::empty())?;
 
             let res = self.send_request(req).await?;
-            if !res.status().is_server_error() && !res.status().is_client_error() {
-                Ok(res.into_body().data().await.unwrap_or(Ok(Bytes::new()))?.to_vec())
-            } else {
-                Err(ResponseError(res.status()).into())
-            }
+            let bytes = hyper::body::to_bytes(res.into_body()).await?;
+            let file = String::from_utf8(bytes.to_vec())?;
+
+            Ok(Document::from(file.as_str()))
         }
     }
 }
@@ -92,11 +89,12 @@ impl Client {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use select::document::Document;
 
     use std::error::Error;
     async fn conn_client(host: &str) -> Result<(), Box<dyn Error>> {
         let mut client = Client::new(host).await?;
-        let res = client.get_bytes("/").await?;
+        let _res = client.query_html("/").await?;
 
         Ok(())
     }
@@ -106,7 +104,6 @@ mod tests {
         conn_client("google.com").await.unwrap();
         conn_client("example.com").await.unwrap();
         conn_client("naver.com").await.unwrap();
-        conn_client("knowhow.or.kr").await.unwrap();
     }
 
     #[tokio::test]
@@ -114,6 +111,5 @@ mod tests {
         conn_client("e-hentai.org").await.unwrap();
         conn_client("hitomi.la").await.unwrap();
         conn_client("xvideos.com").await.unwrap();
-        conn_client("manatoki.net").await.unwrap();
     }
 }
